Reversed Pythia Input Changes and Adjusted FACET implementation
I reversed the changes back from "IN" to "CMND".
I also changed the corner points array of FACET, as the FACET detector does not cover eta up to infinity like FASER does. Its shape seems to be that of a hollow cylinder similar to AL3X, so i adjusted the corner points to be similar to AL3X. If the covered minimal polar angle is 1mrad, the inner radius of the hollow cylinder should be approximately101*tan( 1E-3 ) = 0.1.